Move the code from __init__ to its module
