+++ quran_tafseer_bot/settings.py
@@ -5,4 +5,4 @@ load_dotenv(find_dotenv())
 
 ROOT_PATH=os.path.dirname(os.path.abspath(__file__))
 BOT_TOKEN=os.getenv('QURAN_TAFSEER_BOT_TOKEN')
-TEMAPLTE_PATH=os.path.join(ROOT_PATH, 'templates')
+TEMPLATE_PATH=os.path.join(ROOT_PATH, 'templates')
