+++ quran_tafseer_bot/templates/setbook_invalid.html
@@ -1,0 +1,1 @@
+Please provide the book ID after the setbook command, like /setbook 2
