+++ quran_tafseer_bot/templates/tafseer_invalid.html
@@ -1,0 +1,1 @@
+Please provide the chapter number and the verse number the tafseer command, like /tafseer 1 1
